requirements.txt
python-dotenv==1.0.1
build>=1.2.2.post1
pydantic==v2.11.7
typer==0.16.0
pytz==2025.2
requests==2.32.5
setuptools==78.1.1
wheel==0.43.0
pytest==8.0.1
pytest-cov==4.1.0
pytest-env==1.1.3
ruff==v0.4.4
freezegun==1.2.1
